* release/1.1.1:
  dynamic meta, ,meta translations and focus on ES country, clean old titles
  Cache and Redis
  Add translation support, english default
  Update permissions
  Improve amenities permissions

+++ app/Http/Controllers/Admin/AmenityController.php
@@ -12,6 +12,8 @@ class AmenityController extends Controller
 {
     public function index()
     {
+        $this->authorize('view amenities');
+        
         $Amenity = Amenity::latest()->paginate(15);
         $communities = Community::all();
 
@@ -20,7 +22,8 @@ class AmenityController extends Controller
 
     public function create()
     {
-        // Return the create view
+        $this->authorize('create amenities');
+        
         return view('admin.amenities.create');
     }
 
@@ -45,27 +48,27 @@ class AmenityController extends Controller
             'description' => $request->description,
         ]);
 
-        return redirect()->route('amenity.index')->with('success', 'Amenity created successfully.');
+        return redirect()->route('amenity.index')->with('success', __('success.created', ['item' => __('Amenities')]));
     }
 
     public function show(Amenity $amenity)
     {
-        // Show a single Amenity
+        $this->authorize('view amenities');
+        
         return view('admin.amenity.show', compact('amenity'));
     }
 
     public function edit($id)
     {
+        $this->authorize('edit amenities');
+        
         $amenity = Amenity::findOrFail($id);
 
-        // Show the edit form
         return view('admin.amenities.edit', compact('amenity'));
     }
 
     public function update(Request $request, $id)
     {
-        \Log::info($request->all());
-        // dd($id);
         // Validate form data
         $request->validate([
             'name' => 'required|string|max:255',
@@ -92,11 +95,13 @@ class AmenityController extends Controller
             $amenity->communities()->sync($request->community_ids);
         }
 
-        return redirect()->route('amenity.index')->with('success', 'Amenity updated successfully.');
+        return redirect()->route('amenity.index')->with('success', __('success.updated', ['item' => __('Amenities')]));
     }
 
     public function destroy($id)
     {
+        $this->authorize('delete amenities');
+        
         $Amenity = Amenity::findOrFail($id);
         Storage::disk('public')->delete($Amenity->logo);
         $Amenity->delete();
@@ -105,6 +110,6 @@ class AmenityController extends Controller
             return response('', 200);
         }
 
-        return redirect()->route('amenity.index')->with('success', 'Amenity deleted successfully.');
+        return redirect()->route('amenity.index')->with('success', __('success.deleted', ['item' => __('Amenities')]));
     }
 }
