+++ app/Http/Controllers/Admin/PermissionController.php
@@ -11,16 +11,22 @@ class PermissionController extends Controller
 {
     public function index()
     {
+        $this->authorize('manage permissions');
+        
         $permissions = Permission::latest()->paginate(15);
         return view('admin.permission.index', compact('permissions'));
     }
     public function create()
     {
+        $this->authorize('manage permissions');
+        
         $data['permissions'] = Permission::get();
         return view('admin.permission.create', $data);
     }
     public function store(Request $request)
     {
+        $this->authorize('manage permissions');
+        
         $this->validate($request, [
             'name' => 'required',
         ]);
@@ -30,17 +36,21 @@ class PermissionController extends Controller
         Permission::create($input);
 
         return redirect()->route('permission.index')
-        ->with('success', 'Permission updated successfully');
+        ->with('success', __('success.updated', ['item' => __('Permissions')]));
     }
 
     public function edit($id)
     {
+        $this->authorize('manage permissions');
+        
         $permission = Permission::find($id);
         return view('admin.permission.edit', compact('permission'));
     }
 
     public function update(Request $request, $id)
     {
+        $this->authorize('manage permissions');
+        
         $this->validate($request, [
             'name' => 'required',
         ]);
@@ -52,10 +62,12 @@ class PermissionController extends Controller
         $user->update($input);
 
         return redirect()->route('permission.index')
-            ->with('success', 'Permission updated successfully');
+            ->with('success', __('success.updated', ['item' => __('Permissions')]));
     }
     public function destroy($id)
     {
+        $this->authorize('manage permissions');
+        
         $permission = Permission::findOrFail($id);
         $permission->delete();
 
@@ -63,7 +75,7 @@ class PermissionController extends Controller
             return response('', 200);
         }
 
-        return redirect()->route('permission.index')->with('success', 'Permission deleted successfully.');
+        return redirect()->route('permission.index')->with('success', __('success.deleted', ['item' => __('Permissions')]));
     }
 
 }
